sm8250-common: bootctrl: Remove Android.mk rules for building bootctrl.
All the targets that need the bootctrl.$(TARGET_BOARD_PLATFORM) in this
dir should build the module with Android.bp, which is also the desired
way to install the recovery variant of bootctrl. [1] has an example
change for crosshatch.

We also no longer need BOARD_IS_AUTOMOTIVE, as the module must be
explicitly included now, in order to be built with namespace-scoped
Android.bp.

[1] https://android-review.googlesource.com/c/device/google/crosshatch/+/958004

Bug: 34254109
Test: TreeHugger
Change-Id: Ib110508065f47a742acd92e03ea42901e8002e4f