nio: fix speaker-dmic-endfire mixer path
nio has three mics:
  - bottom: ADC0
  - top: ADC1
  - back: ADC2

`speaker-dmic-endfire` is a dual-mic usecase, but currently configured
to enable ADC1 (top) only, which is inconsistent with the topology
defined in ACDB, so no audio frame will be captured.

This patch replaces the usecase with `camcorder-mic-bf`, which is
properly defined for dual-mic. ADC1 is the left channel and ADC0 is the
right channel.

Change-Id: Ib7a843ab69dc7dea8c3c7906cea34c50349d5cdb